botCommands.py: new file and migrate code to add events
The botCommands file will hold source code for functions used by the
bot to collect information and execute code. Such as eventAdd(),
which parses the user's arguments instead of in the core bot file.
This will help reduce the clutter in the core bot file, as well as
provide better factorization.

botCommands.py:
+ eventAdd()

gogobeBotCore.py:
~ event [command]
  - event add [command]