events.py: add functions to format the bot's table of events output
formatNumber() formats numbers to fit nicely in the bot's table,
regardless if there are four digits or one (the table is only
designed for up to four digits in the event number or the SQL ID.

getEvents() gets the given number of most recent events of a given
title. e.g., the 5 most recent CDLs.

+ formatNumber()
+ getEvents()