gogobeBotCore.py: remove status command and sqlpyte
The status command was used for testing if the database and its
functions were working properly. While useful for testing, a user's
status has no purpose for the CDL, so it's getting remove. It will
be replaced with something else more useful.

The sqlpyte imports were also removed as they are no longer needed in
this script file.

- status [command]