metaDB.py: add func to create standings table
This creates a table to hold the overall stats in the CDL, such as
medal totals, score total, etc. This may be made obsolete later in
development by simply calculating totals from individual results, but
until then, this will allow for the bot to take shape and return info
to users.

+ createStandingsTable()